Improve error catch from juction code path

+++ lib/spack/llnl/util/symlink.py
@@ -30,9 +30,15 @@ def symlink(real_path, link_path):
         try:
             # Try to use junctions
             _win32_junction(real_path, link_path)
-        except OSError:
-            # If all else fails, fall back to copying files
-            shutil.copyfile(real_path, link_path)
+        except OSError as e:
+            if e.errno == errno.EEXIST:
+                # we manually raised this error, we know shutil will fail the copy
+                # due to a same file error
+                # skip and report
+                raise e
+            else:
+                # If all else fails, fall back to copying files
+                shutil.copyfile(real_path, link_path)
 
 
 def islink(path):
